pyproject.toml
[project]
name = "ocr-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.5.1",
    "httpx[socks]>=0.28.1",
    "onnxruntime>=1.24.1",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.1.1",
    "rapidocr>=3.6.0",
]
